requirements.txt
colorama>=0.4.6
groq>=0.4.0
python-dotenv>=1.0.0
requests>=2.31.0
psycopg2-binary>=2.9.9
dspy-ai>=2.4.6
jsonschema>=4.18.0
argparse>=1.4.0 